Plotting graphs between variables.
YouTube link - https://www.youtube.com/watch?v=aTpNMvRZr-o

The objective is to create a package in python to simplify the task of data exploration. It contains a function that returns separate lists of categorical and numerical variables present in the data set. It returns histogram and boxplot for numerical variables and bar graph for categorical variables.